requirements.txt
streamlit
google.generativeai
time